refactor: migrate IBL baker shader from raw GLSL to ShaderLab format
- Consolidate 4 shader files (vertex.ts, ibl_frag.ts, hammersley.ts,
  importanceSampling.ts) into single ibl-baker.ts ShaderLab source
- Replace attribute/varying/uniform with struct Attributes/Varyings
- Replace #include <common> with #include "Common/Common.glsl"
- No render state declarations in shader (keep engine defaults)
- Update IBLBaker.ts to use new single source import
