--- packages/baker/src/shader/hammersley.ts
@@ -1,44 +1,0 @@
-export default `
-// https://learnopengl.com/PBR/IBL/Specular-IBL
-// Hammersley
-float radicalInverse_VdC(uint bits) 
-{
-    bits = (bits << 16u) | (bits >> 16u);
-    bits = ((bits & 0x55555555u) << 1u) | ((bits & 0xAAAAAAAAu) >> 1u);
-    bits = ((bits & 0x33333333u) << 2u) | ((bits & 0xCCCCCCCCu) >> 2u);
-    bits = ((bits & 0x0F0F0F0Fu) << 4u) | ((bits & 0xF0F0F0F0u) >> 4u);
-    bits = ((bits & 0x00FF00FFu) << 8u) | ((bits & 0xFF00FF00u) >> 8u);
-    return float(bits) * 2.3283064365386963e-10; // / 0x100000000
-}
-
-vec2 hammersley(uint i, uint N)
-{
-    return vec2(float(i)/float(N), radicalInverse_VdC(i));
-}
-
-// WebGL1 fallback
-float VanDerCorpus(uint n, uint base)
-{
-    float invBase = 1.0 / float(base);
-    float denom   = 1.0;
-    float result  = 0.0;
-
-    for(uint i = 0u; i < 32u; ++i)
-    {
-        if(n > 0u)
-        {
-            denom   = mod(float(n), 2.0);
-            result += denom * invBase;
-            invBase = invBase / 2.0;
-            n       = uint(float(n) / 2.0);
-        }
-    }
-
-    return result;
-}
-// ----------------------------------------------------------------------------
-vec2 HammersleyNoBitOps(uint i, uint N)
-{
-    return vec2(float(i)/float(N), VanDerCorpus(i, 2u));
-}
-`;
