--- packages/baker/src/shader/importanceSampling.ts
@@ -1,26 +1,0 @@
-export default `
-vec3 importanceSampleGGX(vec2 Xi, vec3 N, float roughness)
-{
-    float a = roughness * roughness;
-
-    float phi = 2.0 * PI * Xi.x;
-    // NOTE: (aa-1) == (a-1)(a+1) produces better fp accuracy
-    float cosTheta2 = (1.0 - Xi.y) / (1.0 + (a + 1.0) * ((a - 1.0) * Xi.y));
-    float cosTheta = sqrt(cosTheta2);
-    float sinTheta = sqrt(1.0 - cosTheta2);
-
-    // from spherical coordinates to cartesian coordinates
-    vec3 H;
-    H.x = cos(phi) * sinTheta;
-    H.y = sin(phi) * sinTheta;
-    H.z = cosTheta;
-
-    // from tangent-space vector to world-space sample vector
-    vec3 up        = abs(N.z) < 0.999 ? vec3(0.0, 0.0, 1.0) : vec3(1.0, 0.0, 0.0);
-    vec3 tangent   = normalize(cross(up, N));
-    vec3 bitangent = cross(N, tangent);
-
-    vec3 sampleVec = tangent * H.x + bitangent * H.y + N * H.z;
-    return normalize(sampleVec);
-}
-`;
